requirements.txt
flask
flask_restx
flask_marshmallow
flask_sqlalchemy
marshmallow-sqlalchemy 